+++ Node_Bakery/nav_data.py
@@ -13,10 +13,10 @@ class NavData:
     ):
         #node_key ranges from 0-63
         self.node_key = np.uint8(node_key)
-        self.vertices = np.asarray(vertices, dtype = np.float64)
-        self.vertex_normals = np.asarray(vertex_normals, dtype = np.float64)
-        self.tri_centers = np.asarray(tri_centers, dtype = np.float64)
-        self.tri_normals = np.asarray(tri_normals, dtype = np.float64)
+        self.vertices = np.asarray(vertices, dtype = np.float32)
+        self.vertex_normals = np.asarray(vertex_normals, dtype = np.float32)
+        self.tri_centers = np.asarray(tri_centers, dtype = np.float32)
+        self.tri_normals = np.asarray(tri_normals, dtype = np.float32)
         self.tri_vertex_indices = np.asarray(tri_vertex_indices, dtype = np.uint16)
         self.tri_connections = np.asarray(
             tri_connections,
